+++ src/main/java/teammates/logic/core/FeedbackResponseRecordLogic.java
@@ -23,7 +23,7 @@ public class FeedbackResponseRecordLogic {
         return instance;
     }
 
-    public void createFeedbackResponseRecord(int count, int timestamp)
+    public void createFeedbackResponseRecord(long count, long timestamp)
             throws InvalidParametersException, EntityAlreadyExistsException {
         frmDb.createEntity(new FeedbackResponseRecordAttributes(count, timestamp));
     }
